+++ fluxdb/shardinject.go
@@ -19,51 +19,57 @@ import (
 	"encoding/gob"
 	"fmt"
 	"io"
-	"time"
 
 	"github.com/dfuse-io/dstore"
+	"github.com/dfuse-io/shutter"
 	"go.uber.org/zap"
 )
 
 type ShardInjector struct {
+	*shutter.Shutter
+
 	shardsStore dstore.Store
 	db          *FluxDB
 }
 
 func NewShardInjector(shardsStore dstore.Store, db *FluxDB) *ShardInjector {
 	return &ShardInjector{
+		Shutter:     shutter.New(),
 		shardsStore: shardsStore,
 		db:          db,
 	}
 }
 
 func (s *ShardInjector) Run() (err error) {
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Minute)
-	defer cancel()
+	ctx, cancelInjector := context.WithCancel(context.Background())
+	s.Shutter.OnTerminating(func(_ error) {
+		cancelInjector()
+	})
 
 	err = s.shardsStore.Walk(ctx, "", "", func(filename string) error {
 		zlog.Info("processing shard file", zap.String("filename", filename))
 
 		reader, err := s.shardsStore.OpenObject(ctx, filename)
 		if err != nil {
-			return fmt.Errorf("opening object from shards store %q: %s", filename, err)
+			return fmt.Errorf("opening object from shards store %q: %w", filename, err)
 		}
 		defer reader.Close()
 
 		requests, err := readWriteRequestsForBatch(reader)
 		if err != nil {
-			return fmt.Errorf("unable to read all write requests in batch %q: %s", filename, err)
+			return fmt.Errorf("unable to read all write requests in batch %q: %w", filename, err)
 		}
 
 		// TODO: make sure the `LAST` is replaced by a shard-aware value.
-		if err := s.db.WriteBatch(context.Background(), requests); err != nil {
-			return fmt.Errorf("write batch %q: %s", filename, err)
+		if err := s.db.WriteBatch(ctx, requests); err != nil {
+			return fmt.Errorf("write batch %q: %w", filename, err)
 		}
 
 		return nil
 	})
+
 	if err != nil {
-		return fmt.Errorf("walking shards store: %s", err)
+		return fmt.Errorf("walking shards store: %w", err)
 	}
 
 	return nil
@@ -80,7 +86,7 @@ func readWriteRequestsForBatch(reader io.Reader) ([]*WriteRequest, error) {
 			return requests, nil
 		}
 		if err != nil {
-			return nil, fmt.Errorf("unable to read WriteRequest: %s", err)
+			return nil, fmt.Errorf("unable to read WriteRequest: %w", err)
 		}
 		requests = append(requests, req)
 	}